+++ test/testproj-1-module.t/run.t
@@ -38,9 +38,6 @@ Check that files were created as expected:
   main.ml
   ounittest.ml
 
-Set diff command to make sure this test also works under macOS
-  $ export MUTAML_DIFF_COMMAND="diff -u"
-
 Set seed and (full) mutation rate as environment variables, for repeatability
   $ export MUTAML_SEED=896745231
   $ export MUTAML_MUT_RATE=100
@@ -438,99 +435,6 @@ And try the --no-diff option without providing an explicit file name:
   The score is below 100%. Use --fail-under to accept a lower score.
   [2]
 
-
---------------------------------------------------------------------------------
-
-
-And try with a different MUTAML_DIFF_COMMAND environment variable:
-
-  $ export MUTAML_DIFF_COMMAND="diff -U 2"
-  $ mutaml-report
-  Attempting to read from mutaml-report.json...
-  
-  Mutaml report summary:
-  ----------------------
-  
-   target                          #mutations      #failed      #timeouts      #passed 
-   -------------------------------------------------------------------------------------
-   lib.ml                                15      80.0%   12     0.0%    0    20.0%    3
-   =====================================================================================
-  
-  Mutation programs passing the test suite:
-  -----------------------------------------
-  
-  Mutation "lib.ml-mutant6" passed (see "_mutations/lib.ml-mutant6.output"):
-  
-  --- lib.ml
-  +++ lib.ml-mutant6
-  @@ -22,5 +22,5 @@
-   let pi total =
-     let rec loop n inside =
-  -    if n = 0 then
-  +    if n = 1 then
-         4. *. (float_of_int inside /. float_of_int total)
-       else
-  
-  ---------------------------------------------------------------------------
-  
-  Mutation "lib.ml-mutant8" passed (see "_mutations/lib.ml-mutant8.output"):
-  
-  --- lib.ml
-  +++ lib.ml-mutant8
-  @@ -27,5 +27,5 @@
-         let x = 1.0 -. Random.float 2.0 in
-         let y = 1.0 -. Random.float 2.0 in
-  -      if x *. x +. y *. y <= 1.
-  +      if ((x *. x) +. (y *. y)) < 1.
-         then loop (n-1) (inside+1)
-         else loop (n-1) (inside)
-  
-  ---------------------------------------------------------------------------
-  
-  Mutation "lib.ml-mutant14" passed (see "_mutations/lib.ml-mutant14.output"):
-  
-  --- lib.ml
-  +++ lib.ml-mutant14
-  @@ -31,3 +31,3 @@
-         else loop (n-1) (inside)
-     in
-  -  loop total 0
-  +  loop total 1
-  
-  ---------------------------------------------------------------------------
-  
-  Mutation score: 80.0% (15 mutations: 12 failed, 0 timed out, 3 passed)
-  The score is below 100%. Use --fail-under to accept a lower score.
-  [2]
-
-
-Also check that MUTAML_DIFF_COMMAND doesn't affect --no-diff:
-
-  $ mutaml-report --no-diff
-  Attempting to read from mutaml-report.json...
-  
-  Mutaml report summary:
-  ----------------------
-  
-   target                          #mutations      #failed      #timeouts      #passed 
-   -------------------------------------------------------------------------------------
-   lib.ml                                15      80.0%   12     0.0%    0    20.0%    3
-   =====================================================================================
-  
-  Mutation programs passing the test suite:
-  -----------------------------------------
-  
-  Mutation "lib.ml-mutant6" passed (see "_mutations/lib.ml-mutant6.output")
-  Mutation "lib.ml-mutant8" passed (see "_mutations/lib.ml-mutant8.output")
-  Mutation "lib.ml-mutant14" passed (see "_mutations/lib.ml-mutant14.output")
-  Mutation score: 80.0% (15 mutations: 12 failed, 0 timed out, 3 passed)
-  The score is below 100%. Use --fail-under to accept a lower score.
-  [2]
-
-
-Now clean-up MUTAML_DIFF_COMMAND again to default again
-
-  $ export MUTAML_DIFF_COMMAND="diff -u"
 
 --------------------------------------------------------------------------------